extract model changes to module Differ::Model
