+++ lib_static/plugins/acts_as_journalized/lib/acts_as_journalized.rb
@@ -52,7 +52,7 @@ require "cause_of_change"
 module Acts
 end
 
-Dir[File.expand_path("acts/journalized/*.rb", __dir__)].sort.each { |f| require f }
+Dir[File.expand_path("acts/journalized/{,*/}*.rb", __dir__)].each { |f| require f }
 
 module Acts
   module Journalized
