+++ spec/lib/acts_as_journalized/journable_differ_spec.rb
@@ -31,83 +31,6 @@
 require "spec_helper"
 
 RSpec.describe Acts::Journalized::JournableDiffer do
-  describe ".changes" do
-    context "when the objects are work packages" do
-      let(:original) do
-        build_stubbed(:work_package,
-                      subject: "The original work package title",
-                      description: "The description\n",
-                      assigned_to_id: nil,
-                      schedule_manually: false,
-                      ignore_non_working_days: true,
-                      estimated_hours: 1)
-      end
-      let(:changed) do
-        build_stubbed(:work_package,
-                      subject: "The changed work package title",
-                      description: "The description\r\n",
-                      priority: original.priority,
-                      type: original.type,
-                      project: original.project,
-                      schedule_manually: nil,
-                      ignore_non_working_days: false,
-                      estimated_hours: nil)
-      end
-
-      it "returns the changes" do
-        expect(described_class.changes(original, changed))
-          .to eq(
-            "subject" => [original.subject, changed.subject],
-            "author_id" => [original.author_id, changed.author_id],
-            "status_id" => [original.status_id, changed.status_id],
-            "schedule_manually" => [false, nil],
-            "ignore_non_working_days" => [true, false],
-            "estimated_hours" => [1.0, nil]
-          )
-      end
-    end
-
-    context "when the objects are WorkPackageJournal" do
-      let(:original) do
-        build_stubbed(:journal_work_package_journal,
-                      subject: "The original work package title",
-                      description: nil,
-                      priority_id: 5,
-                      type_id: 89,
-                      project_id: 12,
-                      status_id: 45,
-                      schedule_manually: nil,
-                      ignore_non_working_days: false,
-                      estimated_hours: nil)
-      end
-      let(:changed) do
-        build_stubbed(:journal_work_package_journal,
-                      subject: "The changed work package title",
-                      description: "",
-                      priority_id: original.priority_id,
-                      type_id: original.type_id,
-                      project_id: original.project_id,
-                      status_id: original.status_id + 12,
-                      schedule_manually: false,
-                      ignore_non_working_days: true,
-                      estimated_hours: 1)
-      end
-
-      it "returns the changes" do
-        # The description field changes from nil to '', but we want filter those transitions out,
-        # hence the expected hash does not contain the description related change.
-        expect(described_class.changes(original, changed))
-          .to eq(
-            "subject" => [original.subject, changed.subject],
-            "status_id" => [original.status_id, changed.status_id],
-            "schedule_manually" => [nil, false],
-            "ignore_non_working_days" => [false, true],
-            "estimated_hours" => [nil, 1.0]
-          )
-      end
-    end
-  end
-
   describe ".association_changes" do
     context "when the objects are work packages" do
       let(:original) do
